[vkgen] Use correct types for 64-bit types
When I wrote vkgen's stdint.h, qfcc didn't support long or ulong, and
with the common init check, it now matters that the correct types are
used. It turns out some of my structs may not have been initialized
correctly, but with the next commits (long and ulong for cexpr), they
should be but do seem to be working at least.

+++ libs/video/renderer/vulkan/vkgen/stdint.h
@@ -1,13 +1,13 @@
 #ifndef __vkgen_stdint_h
 #define __vkgen_stdint_h
-typedef int uint8_t;
-typedef int uint16_t;
-typedef int uint32_t;
-typedef int uint64_t;
+typedef unsigned uint8_t;
+typedef unsigned uint16_t;
+typedef unsigned uint32_t;
+typedef unsigned long uint64_t;
 typedef int int8_t;
 typedef int int16_t;
 typedef int int32_t;
-typedef int int64_t;
+typedef long int64_t;
 typedef int size_t;
 typedef struct char { int x; } char;//FIXME add char to qfcc
 #endif//__vkgen_stdint_h
